Les fonctions d'affichage dans le fichier main
